cf-authenticated-origin-cert-update.sh update in 123.09beta01
- additional check added to /usr/local/src/centminmod/tools/cf-authenticated-origin-cert-update.sh
- to update run cmupdate

+++ tools/cf-authenticated-origin-cert-update.sh
@@ -21,41 +21,46 @@ cfauthorigin_cronsetup() {
 cforigin() {
   mode=$1
   list_origincrt_domains=$(find ${cf_auth_origin_cert_dir} -type f -name 'origin.crt' -exec dirname {} \; 2>&1 | sed -e "s|${cf_auth_origin_cert_dir}/||g")
-  for d in $list_origincrt_domains; do
-    expiry=$(openssl x509 -enddate -noout -in "${cf_auth_origin_cert_dir}/${d}/origin.crt" | cut -d'=' -f2 | awk '{print $2 " " $1 " " $4}')
-    epochExpirydate=$(date -d"${expiry}" +%s)
-    epochToday=$(date +%s)
-    secondsToExpire=$(echo ${epochExpirydate} - ${epochToday} | bc)
-    daysToExpire=$(echo "${secondsToExpire} / 60 / 60 / 24" | bc)
-    echo -e "------------------------------\n$d cloudflare authenticated origin cert expires in $daysToExpire days on $expiry"
-    if [[ "$daysToExpire" -le '180' ]]; then
-      if [[ "$mode" = 'check' ]]; then
-        echo "at ${cf_auth_origin_cert_dir}/${d}/origin.crt"
-      elif [[ "$mode" = 'update' ]]; then
-        echo "updating $d cloudflare authenticated origin cert"
-        echo "at ${cf_auth_origin_cert_dir}/${d}/origin.crt"
-        cp -a "${cf_auth_origin_cert_dir}/${d}/origin.crt" "${cf_auth_origin_cert_dir}/${d}/origin.crt-backup"
-        wget -4 -q -O "${cf_auth_origin_cert_dir}/${d}/origin.crt" "$cf_auth_origin_cert"
-        err=$?
-        if [[ "$err" -eq '0' ]]; then
-          echo "succesfully updated ${cf_auth_origin_cert_dir}/${d}/origin.crt"
-          rm -f "${cf_auth_origin_cert_dir}/${d}/origin.crt-backup"
-          service nginx reload >/dev/null 2>&1
-          expiry=$(openssl x509 -enddate -noout -in "${cf_auth_origin_cert_dir}/${d}/origin.crt" | cut -d'=' -f2 | awk '{print $2 " " $1 " " $4}')
-          epochExpirydate=$(date -d"${expiry}" +%s)
-          epochToday=$(date +%s)
-          secondsToExpire=$(echo ${epochExpirydate} - ${epochToday} | bc)
-          daysToExpire=$(echo "${secondsToExpire} / 60 / 60 / 24" | bc)
-          echo -e "$d cloudflare authenticated origin cert now expires in $daysToExpire days on $expiry"
-        else
-          echo "failed to update ${cf_auth_origin_cert_dir}/${d}/origin.crt"
-          echo "restoring previous backup"
-          \cp -af "${cf_auth_origin_cert_dir}/${d}/origin.crt-backup" "${cf_auth_origin_cert_dir}/${d}/origin.crt"
-          service nginx reload >/dev/null 2>&1
+  if [ "$list_origincrt_domains" ]; then
+    for d in $list_origincrt_domains; do
+      expiry=$(openssl x509 -enddate -noout -in "${cf_auth_origin_cert_dir}/${d}/origin.crt" | cut -d'=' -f2 | awk '{print $2 " " $1 " " $4}')
+      epochExpirydate=$(date -d"${expiry}" +%s)
+      epochToday=$(date +%s)
+      secondsToExpire=$(echo ${epochExpirydate} - ${epochToday} | bc)
+      daysToExpire=$(echo "${secondsToExpire} / 60 / 60 / 24" | bc)
+      echo -e "------------------------------\n$d cloudflare authenticated origin cert expires in $daysToExpire days on $expiry"
+      if [[ "$daysToExpire" -le '180' ]]; then
+        if [[ "$mode" = 'check' ]]; then
+          echo "at ${cf_auth_origin_cert_dir}/${d}/origin.crt"
+        elif [[ "$mode" = 'update' ]]; then
+          echo "updating $d cloudflare authenticated origin cert"
+          echo "at ${cf_auth_origin_cert_dir}/${d}/origin.crt"
+          cp -a "${cf_auth_origin_cert_dir}/${d}/origin.crt" "${cf_auth_origin_cert_dir}/${d}/origin.crt-backup"
+          wget -4 -q -O "${cf_auth_origin_cert_dir}/${d}/origin.crt" "$cf_auth_origin_cert"
+          err=$?
+          if [[ "$err" -eq '0' ]]; then
+            echo "succesfully updated ${cf_auth_origin_cert_dir}/${d}/origin.crt"
+            rm -f "${cf_auth_origin_cert_dir}/${d}/origin.crt-backup"
+            service nginx reload >/dev/null 2>&1
+            expiry=$(openssl x509 -enddate -noout -in "${cf_auth_origin_cert_dir}/${d}/origin.crt" | cut -d'=' -f2 | awk '{print $2 " " $1 " " $4}')
+            epochExpirydate=$(date -d"${expiry}" +%s)
+            epochToday=$(date +%s)
+            secondsToExpire=$(echo ${epochExpirydate} - ${epochToday} | bc)
+            daysToExpire=$(echo "${secondsToExpire} / 60 / 60 / 24" | bc)
+            echo -e "$d cloudflare authenticated origin cert now expires in $daysToExpire days on $expiry"
+          else
+            echo "failed to update ${cf_auth_origin_cert_dir}/${d}/origin.crt"
+            echo "restoring previous backup"
+            \cp -af "${cf_auth_origin_cert_dir}/${d}/origin.crt-backup" "${cf_auth_origin_cert_dir}/${d}/origin.crt"
+            service nginx reload >/dev/null 2>&1
+          fi
         fi
       fi
-    fi
-  done
+    done
+  else
+    echo "no Cloudflare Authenticated Origin Pull Certs to update"
+    echo "at ${cf_auth_origin_cert_dir}"
+  fi
 }
 
 case "$1" in
